Restructure assets folder
