+++ idb/src/assets/images/splash-screen/pic1.jpg
