+++ idb/src/assets/images/splash-screen/pic2.jpg
